Double-encode serif and sans-serif alphabets
This is a sans font, so math alphabets should be all sans even the ones
that are explicitly serif in Unicode. This commit deletes the serif
glyphs and double-encodes the sans-serif glyphs to provides glyphs for
both alphabets.

There are still a few serif glyphs, mostly for ranges that does not have
sans-serif Unicode code points. These will be handled in followup
commits.

Related to https://github.com/notofonts/math/issues/54